requirements.txt
openai
python-dotenv
langchain
langchain-openai
langchain-community
faiss-cpu
pypdf
python-docx
docx2txt
beautifulsoup4
upstash-redis